second update

+++ twoGirlsOnlineShop/Controllers/UserPanelController.cs
@@ -6,58 +6,27 @@ using System.Security.Claims;
 using TwoGirls.Core.Convertors;
 using TwoGirls.Core.DTOs;
 using TwoGirls.Core.Generator;
+using TwoGirls.Core.Security;
 using TwoGirls.Core.Services.Interfaces;
 using TwoGirls.DataLayer.Entities;
 namespace twoGirlsOnlineShop.Controllers
 {
     public class UserPanelController : Controller
     {
-        IUserService _userService;
-        private IWebHostEnvironment _hostingEnvironment;
-        public UserPanelController(IUserService userService, IWebHostEnvironment hostingEnvironment)
+       private readonly IUserService _userService;
+       private readonly IOrderService  _orderService;
+       private readonly IWebHostEnvironment _hostingEnvironment;
+        public UserPanelController(IUserService userService, IOrderService orderService,IWebHostEnvironment hostingEnvironment)
         {
             _userService = userService;
-            _hostingEnvironment = hostingEnvironment;   
+            _hostingEnvironment = hostingEnvironment;
+            _orderService = orderService;
         }
         public IActionResult Index()
         {
             return View();
         }
-        [HttpPost]
-        public IActionResult AddCreditCard(CreditCard creditCard)
-        {
-            if (ModelState.IsValid)
-            {
-                int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
-                creditCard.UserId = userId;
-                _userService.AddUserCreditCard(creditCard);
-
-                var creditCards = _userService.GetUserCreditCards(userId);
-                string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
-                return Json(new { success = true, html = partialView });
-            }
-            else
-            {
-                string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_AddUserCreditCard", creditCard);
-                return Json(new { success = false,  html = partialView });
-            }
-        }
-        [HttpPost]
-        public IActionResult RemoveCreditCard(int creditCardId)
-        {
-            IEnumerable<CreditCard> creditCards;
-            string partialView;
-            int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
-            if (_userService.RemoveUserCreditCard(userId, creditCardId))
-            {
-                creditCards = _userService.GetUserCreditCards(userId);
-                partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
-                return Json(new { success = true, html = partialView });
-            }
-            creditCards = _userService.GetUserCreditCards(userId);
-            partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
-            return Json(new { success = false, html = partialView });
-        }
+        #region User Address
         [HttpPost]
         public IActionResult RemoveAddress(int addressId)
         {
@@ -79,9 +48,11 @@ namespace twoGirlsOnlineShop.Controllers
         {
             if (ModelState.IsValid)
             {
-                _userService.AddUserAddress(address);
                 int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+                address.UserId = userId;
+                _userService.AddUserAddress(address);
                 var addresses = _userService.GetUserAddresses(userId);
+                ViewBag.showRadio = true;
                 string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserAddresses", addresses);
                 return Json(new { success = true, html = partialView });
             }
@@ -91,18 +62,62 @@ namespace twoGirlsOnlineShop.Controllers
                 return Json(new { success = false, html = partialView });
             }
         }
+        #endregion
 
+        #region User CreditCard
         [HttpPost]
-        public IActionResult ChangeAvatar(UserAvatarViewModel userAvatar) {
-
+        public IActionResult AddCreditCard(CreditCard creditCard)
+        {
             if (ModelState.IsValid)
             {
                 int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
-                if(userAvatar.ImagePath != "/image/user-avatar/Default-Avatar.png")
-                {
-                    string oldfilePath = Path.Combine(_hostingEnvironment.WebRootPath,userAvatar.ImagePath);
+                creditCard.UserId = userId;
+                _userService.AddUserCreditCard(creditCard);
 
-                    if(System.IO.File.Exists(oldfilePath))
+                var creditCards = _userService.GetUserCreditCards(userId);
+                string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
+                return Json(new { success = true, html = partialView });
+            }
+            else
+            {
+                string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_AddUserCreditCard", creditCard);
+                return Json(new { success = false, html = partialView });
+            }
+        }
+        [HttpPost]
+        public IActionResult RemoveCreditCard(int creditCardId)
+        {
+            IEnumerable<CreditCard> creditCards;
+            string partialView;
+            int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (_userService.RemoveUserCreditCard(userId, creditCardId))
+            {
+                creditCards = _userService.GetUserCreditCards(userId);
+                partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
+                return Json(new { success = true, html = partialView });
+            }
+            creditCards = _userService.GetUserCreditCards(userId);
+            partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserCreditCards", creditCards);
+            return Json(new { success = false, html = partialView });
+        }
+        #endregion
+
+        #region User Avatar
+
+        [HttpPost]
+        public IActionResult ChangeAvatar(UserAvatarViewModel userAvatar)
+        {
+
+            if (ModelState.IsValid && userAvatar.NewImage.IsImage())
+            {
+                int userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+                if (userAvatar.ImagePath != "/image/user-avatar/Default-Avatar.png")
+                {
+                    string webRootPath = _hostingEnvironment.WebRootPath;
+                    string path = userAvatar.ImagePath.TrimStart('/'); // Remove the leading forward slash
+                    string oldfilePath = Path.Combine(webRootPath, path);
+
+                    if (System.IO.File.Exists(oldfilePath))
                     {
                         System.IO.File.Delete(oldfilePath);
                     }
@@ -116,7 +131,7 @@ namespace twoGirlsOnlineShop.Controllers
                     userAvatar.NewImage.CopyTo(fileStream);
                 }
                 var imagePath = "/image/user-avatar/" + uniqueFileName;
-                userAvatar = _userService.ChangeAvatar(userId,imagePath);
+                userAvatar = _userService.ChangeAvatar(userId, imagePath);
                 string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_UserAvatar", userAvatar);
                 return Json(new { success = true, html = partialView });
             }
@@ -126,7 +141,9 @@ namespace twoGirlsOnlineShop.Controllers
                 return Json(new { success = false, html = partialView });
             }
         }
+        #endregion
 
+        #region User Info
         [HttpPost]
         public async Task<IActionResult> EditUser(EditProfileViewModel editProfile)
         {
@@ -158,8 +175,10 @@ namespace twoGirlsOnlineShop.Controllers
                 return Json(new { success = false, html = partialView });
             }
         }
-        #region Change Password
 
+        #endregion
+
+        #region Change Password
         [HttpPost]
         public IActionResult ChangePassword(ChangePasswordViewModel changePassword)
         {
@@ -177,6 +196,26 @@ namespace twoGirlsOnlineShop.Controllers
                 string partialView = RenderViewToString.RenderRazorViewToString(this, "UserPanel/_ChangePasswordForm", changePassword);
                 return Json(new { success = false, isIncorrect = false, html = partialView });
             }
+        }
+        #endregion
+
+        #region User Orders
+        public IActionResult OrderList()
+        {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return NotFound();
+            }
+            int userId = int.Parse(HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+            var user = _userService.GetUserById(userId);
+            ViewData["UserAvatarViewModel"] = new UserAvatarViewModel()
+            {
+                ImagePath = user.ImagePath,
+                FullName = user.FirstName + " " + user.LastName,
+                Email = user.Email
+            };
+            ViewData["OrderList"] = _orderService.GetAllOrdersByUserId(userId);
+            return View();
         }
         #endregion
     }
